cloud-server-side/pyproject.toml
[project]
name = "cloud-server-side"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.14",
    "huggingface-hub>=1.3.2",
    "langchain>=1.2.4",
    "numpy>=2.4.1",
    "onnxruntime>=1.23.2",
    "tokenizers>=0.22.2",
]
